refactoring :: jwt middleware
