+++ src/types/index.ts
@@ -1,14 +1,21 @@
 import { Request } from 'express';
-import { JwtPayloadData } from './auth';
 import { ParsedQs } from 'qs';
 
 export interface BasicResponse {
   message: string;
 }
 
+export interface JwtPayloadData {
+  id: string;
+  sub: string;
+  type: 'access' | 'refresh';
+  iat: number;
+}
+
 export interface AuthenticatedRequest<Params = Record<string, never>, Query = ParsedQs, Body = Record<string, never>>
   extends Request<Params, any, Body, Query> {
   payload?: JwtPayloadData;
+  userId?: number;
 }
 
 export const REDIS_KEY = {
